backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
jinja2
python-multipart
python-dotenv
sqlmodel
